% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/razip-package.R
\docType{package}
\name{razip-package}
\alias{razip}
\alias{razip-package}
\title{razip: Efficient Random Access to R Objects Stored in Large Zip
    Files}
\description{
Provides efficient random access to the contents of
    large zip files by storing cacheing zip offsets in memory. Contents
    can then be read directly to memory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/razip}
  \item Report bugs at \url{https://github.com/jefferis/razip/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
