% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/razip.R
\name{RAZip}
\alias{RAZip}
\title{R6 class implementing fast random access to zip files.}
\description{
R6 class implementing fast random access to zip files.

R6 class implementing fast random access to zip files.
}
\examples{
\donttest{
smp=RAZip$new(system.file("sampledata/Cell07PNs-rds.zip", package = 'razip'))
smp
}
\dontrun{
raz=RAZip$new("~/Desktop/flywire_neurons_flow_FlyWireqs.zip")
raz
zl=raz$ziplist()
bench::mark(s1=raz$get(sample(zl$filename, 1)), check = F)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{zipfile}}{the path to the zip file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RAZip$new()}}
\item \href{#method-getbytes}{\code{RAZip$getbytes()}}
\item \href{#method-get}{\code{RAZip$get()}}
\item \href{#method-mget}{\code{RAZip$mget()}}
\item \href{#method-ziplist}{\code{RAZip$ziplist()}}
\item \href{#method-clone}{\code{RAZip$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise an RAZip object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAZip$new(zipfile, unserialize = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zipfile}}{path to the zip file}

\item{\code{unserialize}}{Optional function to be applied to raw bytes from the
zip file. When missing R's \code{unserialize} or \code{qs::deserialize}
is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getbytes"></a>}}
\if{latex}{\out{\hypertarget{method-getbytes}{}}}
\subsection{Method \code{getbytes()}}{
read raw bytes for a zip entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAZip$getbytes(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{A file inside the zip to read.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
read an R object from a zip entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAZip$get(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{A file inside the zip to read.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mget"></a>}}
\if{latex}{\out{\hypertarget{method-mget}{}}}
\subsection{Method \code{mget()}}{
read multiple R objects from a zip file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAZip$mget(f, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{One or more files inside the zip to read}

\item{\code{...}}{additional arguments passed to \code{pbapply::\link{pbsapply}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named \code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ziplist"></a>}}
\if{latex}{\out{\hypertarget{method-ziplist}{}}}
\subsection{Method \code{ziplist()}}{
Return the zip info from \code{zip::zip_list}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAZip$ziplist()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named \code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RAZip$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
