% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip2.R
\name{read_zip_entry}
\alias{read_zip_entry}
\title{Read data from one zip file entry}
\usage{
read_zip_entry(zip, f, zl = NULL, unserialize = FALSE)
}
\arguments{
\item{zip}{The path to the zip file or an open connection}

\item{f}{The file to read from inside the zip}

\item{zl}{A zip list as produced by}

\item{unserialize}{Whether to unserialize the data (from rds or qs format).
Otherwise raw bytes are returned.}
}
\value{
A \code{raw} array, which may have been unserialized.
}
\description{
Read data from one zip file entry
}
\details{
When \code{unserialize=TRUE} the data must a) have been written to
  disk with \code{\link{saveRDS}} using gzip, bzip2 or xz compression or
  \code{qs::\link[qs]{qsave}} AND b) zip must have been instructed \emph{not}
  to further compress the data (which would be pointless). This is the
  behaviour of
  \href{https://natverse.org/nat/reference/write.neurons.html}{\code{write.neurons}}
   in the \href{https://natverse.org/nat/}{nat package}.
}
