% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/razip-package.R
\docType{package}
\name{razip-package}
\alias{razip}
\alias{razip-package}
\title{razip: Efficient Random Access to R Objects Stored in Large Zip Files
    Files}
\description{
Provides efficient random access to the contents of large zip files
    by cacheing zip files offsets in memory. Contents can then be read directly
    to memory, optionally unserialising. The main intended use case is the
    storage of collections of tens of thousands of serialised R objects into
    single zip files that may be GB in size while still allowing efficient
    (order 5ms) read access times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/razip}
  \item Report bugs at \url{https://github.com/jefferis/razip/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
